Refactor GoBoadManager.class Fixes #44
